Primera version de la funcionalidad crear inmueble terminada

+++ src/main/java/com/formation/microservice_inmueble/repository/InmuebleDao.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface InmuebleDao extends CrudRepository<Inmueble, Integer> {
+public interface InmuebleDao extends CrudRepository<Inmueble, Long> {
     public Optional<Inmueble> findByCodigoInmueble(int codigoInmueble);
 }
